+++ apps/asteroids/src/app.rs
@@ -256,7 +256,7 @@ impl HostedGame for Asteroids {
         menu::menus()
     }
 
-    fn tick(&mut self, _gpu: &mut Gpu, _tick_dt: f32) {
+    fn tick(&mut self, _gpu: &mut Gpu, _tick_dt: f64) {
         self.game.tick();
     }